Prefix and Suffix Search
Design a special dictionary which has some words and allows you to search the words in it by a prefix and a suffix.

Implement the WordFilter class:

    WordFilter(string[] words) Initializes the object with the words in the dictionary.
    f(string prefix, string suffix) Returns the index of the word in the dictionary which has the prefix prefix and the suffix suffix. If there is more than one valid index, return the largest of them. If there is no such word in the dictionary, return -1.

 

Example 1:

Input
["WordFilter", "f"]
[[["apple"]], ["a", "e"]]
Output
[null, 0]

Explanation
WordFilter wordFilter = new WordFilter(["apple"]);
wordFilter.f("a", "e"); // return 0, because the word at index 0 has prefix = "a" and suffix = 'e".

 

Constraints:

    1 <= words.length <= 15000
    1 <= words[i].length <= 10
    1 <= prefix.length, suffix.length <= 10
    words[i], prefix and suffix consist of lower-case English letters only.
    At most 15000 calls will be made to the function f.

   Hide Hint #1  
For a word like "test", consider "#test", "t#test", "st#test", "est#test", "test#test". Then if we have a query like prefix = "te", suffix = "t", we can find it by searching for something we've inserted starting with "t#te".

Idea:

Whenever we have to deal with searching for data using a prefix or a suffix, this naturally points to a trie solution. A trie is a type of data structure that uses a branching tree format where the nodes represent segments of data (usually characters) to make searching by prefix faster and easier.

The difficulty in this case is that we're searching by both prefix and suffix, so we can create two trie structures, one for prefixes and one for suffixes (pTrie, sTrie). Then we can iterate through words and insert() each word into the two tries.

To do so, we'll iterate through the characters of the word, forwards for pTrie and backwards for sTrie, and move from node to node as the word moves from character to character. At each node, we'll update the vals array with the current index. The vals array represents the indices of all the words that run through the current node. Since we're iterating through words in index order, each node's vals array will be sorted in index order, as well.

For our find method, f(), we'll be doing the same in reverse. We'll navigate separately through pTrie with pre and sTrie with suf to find the vals arrays containing the indices of each word that matches those prefixes and suffixes. If at any point a particular trie does not have the next character, we can return -1.

Once we've successfully obtained the two vals arrays (pVals, sVals), we can cross-reference their contents, starting at the end, and look for the largest index that occurs in both. If we find one, we can return it, otherwise we can return -1.